assignment 17: more deleting / backspace

+++ coaster/src/editor/view.rs
@@ -317,8 +317,16 @@ impl View {
     }
 
     fn backspace(&mut self) {
-        self.move_left();
-        self.delete();
+        if self.text_location.line_index != 0
+            || self
+                .text_location
+                .grapheme_index
+                .saturating_sub(self.line_padding + 1)
+                != 0
+        {
+            self.move_text_location(&Direction::Left);
+            self.delete()
+        }
     }
 
     fn delete(&mut self) {
@@ -343,7 +351,7 @@ impl View {
             .map_or(0, Line::grapheme_count);
         let grapheme_delta = new_len.saturating_sub(old_len);
         if grapheme_delta > 0 {
-            self.move_right();
+            self.move_text_location(&Direction::Right)
         }
         self.needs_redrawn = true;
     }
